🐛 [#068] - Fix east-of-UTC date bug, soft-delete validation, and controller/service split 🔍
- 🐛 Fix toIsoDate() in week.ts to read local Date components instead of toISOString() — UTC extraction shifted dates by one day (and caused weeksInYear's while loop to never terminate) for any timezone east of UTC, since local midnight only agrees with its UTC calendar day at or west of UTC
- 🐛 Reject soft-deleted employees in WeeklySummaryRequest validation via Rule::exists()->whereNull('deleted_at'), matching the pattern used by every other employee-scoped request — previously passed validation then crashed with a 404 in the employee() accessor
- 🔨 Extract WeeklySummaryController's business logic into WeeklySummaryService, following the PayPeriodPreviewService/PreviewPayPeriodController pattern mandated by CLAUDE.md's FormRequest/Controller/Service convention
- ✅ Add regression test for soft-deleted employee returning 422 instead of a crash
